extract block providers to a separate file
